Entities and DBContext seperated
