+++ src/Chirp.Core/Enitities.cs
@@ -1,20 +1,3 @@
-using Microsoft.EntityFrameworkCore;
-using System;
-
-public class ChirpDBContext : DbContext
-{
-    public DbSet<Cheep> Cheeps { get; set; }
-    public DbSet<Author> Authors { get; set; }
-    public ChirpDBContext(DbContextOptions<ChirpDBContext> options) : base(options)
-    { 
-    }
-
-    protected override void OnModelCreating(ModelBuilder modelBuilder)
-    {
-        base.OnModelCreating(modelBuilder);
-        modelBuilder.Entity<Cheep>().Property(x => x.CheepId);
-    }
-}
 
 public class Author
 {
