+++ src/Chirp.Infrastructure/ChirpDBContext.cs
@@ -1,0 +1,16 @@
+using Microsoft.EntityFrameworkCore;
+
+public class ChirpDBContext : DbContext
+{
+    public DbSet<Cheep> Cheeps { get; set; }
+    public DbSet<Author> Authors { get; set; }
+    public ChirpDBContext(DbContextOptions<ChirpDBContext> options) : base(options)
+    { 
+    }
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        base.OnModelCreating(modelBuilder);
+        modelBuilder.Entity<Cheep>().Property(x => x.CheepId);
+    }
+}
